requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
pydantic==2.5.3
openai>=1.40.0
pillow==10.2.0
python-dotenv==1.0.0
aiofiles==23.2.1
reportlab==4.0.8
pytest==7.4.4
httpx==0.26.0
redis>=5.0.0
